Moving subprocess to a library instead of core
