+++ runtime/java/src/common/rawlib/RawProcessHandle.java
@@ -1,10 +1,12 @@
-package common;
+package common.rawlib;
 
 import java.io.BufferedReader;
 import java.io.OutputStream;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.List;
+
+import common.IOToken;
 
 /**
  * This class represents the `<code>ProcessHandle</code>` type in Silver for
@@ -15,15 +17,15 @@ import java.util.List;
  * 
  * @author RandomActsOfGrammar
  */
-public class ProcessHandle {
+public class RawProcessHandle {
 
     private Process proc;
 
-    public ProcessHandle(Process p) {
+    public RawProcessHandle(Process p) {
         proc = p;
     }
 
-    public ProcessHandle(List<String> cmd_args) {
+    public RawProcessHandle(List<String> cmd_args) {
         ProcessBuilder pb = new ProcessBuilder(cmd_args);
         pb.redirectOutput(ProcessBuilder.Redirect.PIPE);
         pb.redirectInput(ProcessBuilder.Redirect.PIPE);
@@ -40,14 +42,14 @@ public class ProcessHandle {
     //Used for reading stdout from the process
     private BufferedReader our_stdout = null;
     /*Read a line from stdout of the process*/
-    public StringCatter readLineStdout() {
+    public String readLineStdout() {
         try {
             if (our_stdout == null) {
                 InputStream instream = proc.getInputStream();
                 InputStreamReader r = new InputStreamReader(instream);
                 our_stdout = new BufferedReader(r);
             }
-            return new StringCatter(our_stdout.readLine());
+            return our_stdout.readLine();
         }
         catch (Exception e) {
             throw new RuntimeException(e);
@@ -58,14 +60,14 @@ public class ProcessHandle {
     //Used for reading stderr from the process
     private BufferedReader our_stderr = null;
     /*Read a line from stderr of the process*/
-    public StringCatter readLineStderr() {
+    public String readLineStderr() {
         try {
             if (our_stderr == null) {
                 InputStream instream = proc.getErrorStream();
                 InputStreamReader r = new InputStreamReader(instream);
                 our_stderr = new BufferedReader(r);
             }
-            return new StringCatter(our_stderr.readLine());
+            return our_stderr.readLine();
         }
         catch (Exception e) {
             throw new RuntimeException(e);
@@ -94,6 +96,52 @@ public class ProcessHandle {
         catch (Exception e) {
             throw new RuntimeException(e);
         }
+    }
+
+
+    /*
+      These are the public functions which use the Silver types and
+      call the private functions that actually take care of the
+      details.
+     */
+
+    /**
+     * <pre>IOVal<ProcessHandle> ::= cmd::String args::[String] io::IO</pre>
+     */
+    public static NIOVal spawnProcess(StringCatter cmd, ConsCell args, IOToken io) {
+        List<String> full_cmd = ConsCell.toList(args);
+        full_cmd.add(0, cmd.toString());
+        return io.wrap(new RawProcessHandle(full_cmd));
+    }
+
+    /**
+     * <pre>IO ::= p::ProcessHandle msg::String io::IO</pre>
+     */
+    public IOToken sendToProcess(StringCatter msg, IOToken io) {
+        this.writeString(msg.toString());
+        return io;
+    }
+
+    /**
+     * <pre>IOVal<String> ::= p::ProcessHandle io::IO</pre>
+     */
+    public NIOVal readLineFromProcess(IOToken io) {
+        return io.wrap(new StringCatter(this.readLineStdout()));
+    }
+
+    /**
+     * <pre>IOVal<String> ::= p::ProcessHandle io::IO</pre>
+     */
+    public NIOVal readErrLineFromProcess(IOToken io) {
+        return io.wrap(new StringCatter(this.readLineStderr()));
+    }
+
+    /**
+     * <pre>IO ::= p::ProcessHandle io::IO</pre>
+     */
+    public IOToken waitForProcess(IOToken io) {
+        this.waitOnEnd();
+        return io;
     }
 }
